o Make lltok an abstract type, a pointer to structure instead of a plain   struct.  There are still some memory leaks in the parser --- all   tokens should be deallocated.
o Added messagestream, warningstream and errorstream control flags so
  each type of message can be directed to a different stream:
     messagestreamstdout
     messagestreamstderr
     messagestream <file>

+++ src/Headers/cpperror.h
@@ -16,7 +16,7 @@ cppReader_pedwarnWithLine (cppReader *p_pfile, int p_line, int p_column,
 extern void cppReader_error (cppReader *p_pfile, /*@only@*/ cstring p_msg);
 extern void cppReader_warning (cppReader *p_pfile, /*@only@*/ cstring p_msg);
 
-extern void cppReader_printContainingFiles (cppReader *p_pfile) /*@modifies g_msgstream, p_pfile@*/ ;
+extern void cppReader_printContainingFiles (cppReader *p_pfile) /*@modifies g_warningstream, p_pfile@*/ ;
 extern void cppReader_warningLit (cppReader *p_pfile, /*@observer@*/ cstring p_msg);
 
 extern void cppReader_pedwarn (cppReader *p_pfile, /*@only@*/ cstring p_msg);
@@ -24,7 +24,7 @@ extern void cppReader_pedwarn (cppReader *p_pfile, /*@only@*/ cstring p_msg);
 extern void cppReader_pedwarnLit (cppReader *p_pfile, /*@observer@*/ cstring p_msg);
 
 extern void cppReader_errorWithLine (cppReader *p_pfile,
-				     long p_line, long p_column, 
+				     int p_line, int p_column, 
 				     /*@only@*/ cstring p_msg);
 
 extern void cppReader_perrorWithName (cppReader *p_pfile, cstring p_name);
